simple changes

+++ app/src/main/java/com/medhat/zeal/loyaltyapplication/data/repo/DiscountRepo.kt
@@ -1,6 +1,6 @@
 package com.medhat.zeal.loyaltyapplication.data.repo
 
 interface DiscountRepo {
-    fun saveDiscount(discountValue: Long)
-    fun getDiscount(): Long?
+    fun saveDiscount(discountValue: Float)
+    fun getDiscount(): Float?
 }
